requirements.txt
requests~=2.24.0
pendulum~=2.1.2
discord.py~=1.5.1
loguru~=0.5.1